requirements.txt
annotated-types==0.7.0
anyio==4.4.0
click==8.1.7
fastapi==0.114.2
h11==0.14.0
idna==3.8
passlib==1.7.4
psycopg2==2.9.9
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
uvicorn==0.30.6
